# Conflicts:
#	.gitignore
#	VS_LOAN.Core.Nhibernate/XMLConnect/LOAN_hibernate.cfg.xml
#	VS_LOAN.Core.Repository/EmployeeRepository.cs
#	VS_LOAN.Core.Repository/Interfaces/IEmployeeRepository.cs
#	VS_LOAN.Core.Repository/Interfaces/ITailieuRepository.cs
#	VS_LOAN.Core.Repository/TailieuRepository.cs
#	VS_LOAN.Core.Web/Content/ace-master/assets/css/ace.min.css
#	VS_LOAN.Core.Web/Content/css/custom.css
#	VS_LOAN.Core.Web/Controllers/EmployeeController.cs
#	VS_LOAN.Core.Web/Controllers/QuanLyHoSoController.cs
#	VS_LOAN.Core.Web/Views/Employee/AddNew.cshtml
#	VS_LOAN.Core.Web/Views/Employee/Edit.cshtml
#	revoke_debt.sql

+++ VS_LOAN.Core.Web/Controllers/MCreditController.cs
@@ -165,7 +165,7 @@ namespace VS_LOAN.Core.Web.Controllers
                         HosoId = id,
                         Noidung = model.LastNote,
                         CommentTime = DateTime.Now,
-                        TypeId = NoteType.MCreditTemp
+                        TypeId = (int)NoteType.MCreditTemp
                     };
                    await _rpNote.AddNoteAsync(note);
                 }
@@ -203,7 +203,7 @@ namespace VS_LOAN.Core.Web.Controllers
                     HosoId = model.Id,
                     Noidung = model.LastNote,
                     CommentTime = DateTime.Now,
-                    TypeId = NoteType.MCreditTemp
+                    TypeId = (int)NoteType.MCreditTemp
                 };
                 await _rpNote.AddNoteAsync(note);
             }
